requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=1.10.13
pydantic[email]>=1.10.13
python-dotenv>=1.0.1
sqlalchemy>=2.0.27
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
spotipy>=2.23.0
yt-dlp>=2023.12.30
requests>=2.31.0
python-multipart>=0.0.9
pymysql>=1.1.0